Finished simple Genetic Algorithm
Finished a simple implementation of the genetic algorithm. The mutation takes a value and changes it uniformly on a 20% range. Crossover is uniform. Selection is tournament.

Did a proof of concept experiment with score weights, the score weights found by the GA lost against the original parameters. So the GA probably needs to be improved upon to be feasible.