Tweaked the GA
Did some tweaks to the EA. But it does preform horribly. Only tested on score weights, but score weights do not generalise towards a deeper search tree and the original weights do way better than any I find through the GA. Was a fun try, but will probably abandon this method for now and look for a different AI strategy.